complete authentication flow enhance user greetings
