+++ ios/Runner/Info.plist
@@ -45,5 +45,17 @@
 	<true/>
 	<key>ITSAppUsesNonExemptEncryption</key>
     <false/>
+    <key>CFBundleURLTypes</key>
+    <array>
+    	<dict>
+    		<key>CFBundleTypeRole</key>
+    		<string>Editor</string>
+    		<key>CFBundleURLSchemes</key>
+    		<array>
+    			<!-- Copied from GoogleService-Info.plist key REVERSED_CLIENT_ID -->
+    			<string>com.googleusercontent.apps.544972692002-8p565hv8dt84514ren2l8r6pfd840jef</string>
+    		</array>
+    	</dict>
+    </array>
 </dict>
 </plist>
