+++ lib/counters/actions/update_settings.dart
@@ -3,15 +3,32 @@ import 'package:kcounter/counters/core/action.dart';
 import 'package:kcounter/counters/entities/settings_dto.dart';
 import 'package:localstore/localstore.dart';
 
-class UpdateSettings extends ParamsAction<SettingsDto, void> {
+class UpdateSettingsParams {
+  bool? authenticated;
+  bool? online;
+
+  UpdateSettingsParams({
+    this.authenticated,
+    this.online,
+  });
+}
+
+class UpdateSettings extends ParamsAction<UpdateSettingsParams, void> {
   final Localstore db;
 
   UpdateSettings({required this.db});
 
   @override
-  Future<void> run(SettingsDto params) async {
+  Future<void> run(UpdateSettingsParams params) async {
     final collection = db.collection(settingsKey);
     final record = collection.doc(settingsId);
-    await record.set(params.toMap());
+
+    final storedValue = await record.get();
+    final value = storedValue == null ? SettingsDto.defaultValue() : SettingsDto.from(storedValue);
+    final nextValue = value.copyWith(
+      online: params.online ?? value.online,
+      authenticated: params.authenticated ?? value.authenticated,
+    );
+    await record.set(nextValue.toMap());
   }
 }
