+++ lib/widgets/settings_logout_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:kcounter/counters/core/counter_logger.dart';
-import 'package:kcounter/riverpod_providers.dart';
+import 'package:kcounter/riverpod_providers/riverpod_providers.dart';
 import 'package:kcounter/theme/spacing_constants.dart';
 import 'package:kcounter/widgets/counters_button.dart';
 
@@ -22,7 +22,9 @@ class SettingsLogoutButton extends ConsumerWidget {
           horizontal: CountersSpacing.padding900,
         ),
         onPressed: () async {
+          // TODO confirm with the user that he will loose his online counters (they will be copied over the local device)
           try {
+            // TODO copy remote counters to local counters
             await ref.read(signOutProvider.future);
           } catch (error, stacktrace) {
             CounterLogger.error("While trying to sign out", error, stacktrace);
